requirements.txt
discord.py
spotipy
python-dotenv
uuid
motor
dnspython
PyNaCl==1.3.0
pandas
numpy
dnspython==1.16.0
async-timeout==3.0.1